output/quick: Convert sdf.glsl to use highp rather than lowp for everything
We don't want low precision for SDFs as that loses too much information.

+++ src/output/qtquick/plugin/shaders/sdf.glsl
@@ -29,7 +29,7 @@
 //
 // \return The signed distance from point to the circle. If negative, point is
 //         inside the circle.
-lowp float sdf_circle(in lowp vec2 point, in lowp float radius)
+highp float sdf_circle(in highp vec2 point, in highp float radius)
 {
     return length(point) - radius;
 }
@@ -45,22 +45,22 @@ lowp float sdf_circle(in lowp vec2 point, in lowp float radius)
 //
 // \return The signed distance from point to triangle. If negative, point is
 //         inside the triangle.
-lowp float sdf_triangle(in lowp vec2 point, in lowp vec2 p0, in lowp vec2 p1, in lowp vec2 p2)
+highp float sdf_triangle(in highp vec2 point, in highp vec2 p0, in highp vec2 p1, in highp vec2 p2)
 {
-    lowp vec2 e0 = p1 - p0;
-    lowp vec2 e1 = p2 - p1;
-    lowp vec2 e2 = p0 - p2;
+    highp vec2 e0 = p1 - p0;
+    highp vec2 e1 = p2 - p1;
+    highp vec2 e2 = p0 - p2;
 
-    lowp vec2 v0 = point - p0;
-    lowp vec2 v1 = point - p1;
-    lowp vec2 v2 = point - p2;
+    highp vec2 v0 = point - p0;
+    highp vec2 v1 = point - p1;
+    highp vec2 v2 = point - p2;
 
-    lowp vec2 pq0 = v0 - e0 * clamp( dot(v0, e0) / dot(e0, e0), 0.0, 1.0 );
-    lowp vec2 pq1 = v1 - e1 * clamp( dot(v1, e1) / dot(e1, e1), 0.0, 1.0 );
-    lowp vec2 pq2 = v2 - e2 * clamp( dot(v2, e2) / dot(e2, e2), 0.0, 1.0 );
+    highp vec2 pq0 = v0 - e0 * clamp( dot(v0, e0) / dot(e0, e0), 0.0, 1.0 );
+    highp vec2 pq1 = v1 - e1 * clamp( dot(v1, e1) / dot(e1, e1), 0.0, 1.0 );
+    highp vec2 pq2 = v2 - e2 * clamp( dot(v2, e2) / dot(e2, e2), 0.0, 1.0 );
 
-    lowp float s = sign( e0.x*e2.y - e0.y*e2.x );
-    lowp vec2 d = min(min(vec2(dot(pq0,pq0), s*(v0.x*e0.y-v0.y*e0.x)),
+    highp float s = sign( e0.x*e2.y - e0.y*e2.x );
+    highp vec2 d = min(min(vec2(dot(pq0,pq0), s*(v0.x*e0.y-v0.y*e0.x)),
                           vec2(dot(pq1,pq1), s*(v1.x*e1.y-v1.y*e1.x))),
                           vec2(dot(pq2,pq2), s*(v2.x*e2.y-v2.y*e2.x)));
 
@@ -74,9 +74,9 @@ lowp float sdf_triangle(in lowp vec2 point, in lowp vec2 p0, in lowp vec2 p1, in
 //
 // \return The signed distance from point to rectangle. If negative, point is
 //         inside the rectangle.
-lowp float sdf_rectangle(in lowp vec2 point, in lowp vec2 rect)
+highp float sdf_rectangle(in highp vec2 point, in highp vec2 rect)
 {
-    lowp vec2 d = abs(point) - rect;
+    highp vec2 d = abs(point) - rect;
     return length(max(d, 0.0)) + min(max(d.x, d.y), 0.0);
 }
 
@@ -88,11 +88,11 @@ lowp float sdf_rectangle(in lowp vec2 point, in lowp vec2 rect)
 //
 // \return The signed distance from point to rectangle. If negative, point is
 //         inside the rectangle.
-lowp float sdf_rounded_rectangle(in lowp vec2 point, in lowp vec2 rect, in lowp vec4 radius)
+highp float sdf_rounded_rectangle(in highp vec2 point, in highp vec2 rect, in highp vec4 radius)
 {
     radius.xy = (point.x > 0.0) ? radius.xy : radius.zw;
     radius.x = (point.y > 0.0) ? radius.x : radius.y;
-    lowp vec2 d = abs(point) - rect + radius.x;
+    highp vec2 d = abs(point) - rect + radius.x;
     return min(max(d.x, d.y), 0.0) + length(max(d, 0.0)) - radius.x;
 }
 
@@ -106,7 +106,7 @@ lowp float sdf_rounded_rectangle(in lowp vec2 point, in lowp vec2 rect, in lowp 
 // \param thickness The thickness of the resulting shape.
 //
 // \return The value of sdf modified to an annular shape.
-lowp float sdf_annular(in lowp float sdf, in lowp float thickness)
+highp float sdf_annular(in highp float sdf, in highp float thickness)
 {
     return abs(sdf) - thickness;
 }
@@ -118,7 +118,7 @@ lowp float sdf_annular(in lowp float sdf, in lowp float thickness)
 //
 // \return The union of sdf1 and sdf2, that is, the distance to both sdf1 and
 //         sdf2.
-lowp float sdf_union(in lowp float sdf1, in lowp float sdf2)
+highp float sdf_union(in highp float sdf1, in highp float sdf2)
 {
     return min(sdf1, sdf2);
 }
@@ -129,7 +129,7 @@ lowp float sdf_union(in lowp float sdf1, in lowp float sdf2)
 // \param sdf2 The second sdf shape.
 //
 // \return sdf1 with sdf2 subtracted from it.
-lowp float sdf_subtract(in lowp float sdf1, in lowp float sdf2)
+highp float sdf_subtract(in highp float sdf1, in highp float sdf2)
 {
     return max(sdf1, -sdf2);
 }
@@ -141,7 +141,7 @@ lowp float sdf_subtract(in lowp float sdf1, in lowp float sdf2)
 //
 // \return The intersection between sdf1 and sdf2, that is, the area where both
 //         sdf1 and sdf2 provide the same distance value.
-lowp float sdf_intersect(in lowp float sdf1, in lowp float sdf2)
+highp float sdf_intersect(in highp float sdf1, in highp float sdf2)
 {
     return max(sdf1, sdf2);
 }
@@ -153,9 +153,9 @@ lowp float sdf_intersect(in lowp float sdf1, in lowp float sdf2)
 // \param smoothing The amount of smoothing to apply.
 //
 // \return A smoothed version of the intersect operation.
-lowp float sdf_intersect_smooth(in lowp float sdf1, in lowp float sdf2, in lowp float smoothing)
+highp float sdf_intersect_smooth(in highp float sdf1, in highp float sdf2, in highp float smoothing)
 {
-    lowp float h = clamp(0.5 - 0.5 * (sdf1 - sdf2) / smoothing, 0.0, 1.0);
+    highp float h = clamp(0.5 - 0.5 * (sdf1 - sdf2) / smoothing, 0.0, 1.0);
     return mix(sdf1, sdf2, h) + smoothing * h * (1.0 - h);
 }
 
@@ -167,7 +167,7 @@ lowp float sdf_intersect_smooth(in lowp float sdf1, in lowp float sdf2, in lowp 
 // \return The rounded shape of sdf.
 //         Note that rounding happens by basically selecting an isoline of sdf,
 //         therefore, the resulting shape may be larger than the input shape.
-lowp float sdf_round(in lowp float sdf, in lowp float amount)
+highp float sdf_round(in highp float sdf, in highp float amount)
 {
     return sdf - amount;
 }
@@ -177,7 +177,7 @@ lowp float sdf_round(in lowp float sdf, in lowp float amount)
 // \param sdf The sdf shape to turn into an outline.
 //
 // \return The outline of sdf.
-lowp float sdf_outline(in lowp float sdf)
+highp float sdf_outline(in highp float sdf)
 {
     return abs(sdf);
 }
@@ -191,12 +191,12 @@ lowp float sdf_outline(in lowp float sdf)
 // Since 0 is a point exactly on the outline of an sdf shape, and negative
 // values are inside the shape, this uses a very large positive constant to
 // indicate a value that is really far away from the actual sdf shape.
-const lowp float sdf_null = 99999.0;
+const highp float sdf_null = 99999.0;
 
 // A constant for a default level of smoothing when rendering an sdf.
 //
 // This
-const lowp float sdf_default_smoothing = 0.625;
+const highp float sdf_default_smoothing = 0.625;
 
 // Render an sdf shape alpha-blended onto an existing color.
 //
@@ -206,7 +206,7 @@ const lowp float sdf_default_smoothing = 0.625;
 // \param alpha The alpha to use for blending.
 // \param smoothing The amount of smoothing to apply to the sdf.
 //
-lowp vec4 sdf_render(in lowp float sdf, in lowp vec4 sourceColor, in lowp vec4 sdfColor, in lowp float alpha, in lowp float smoothing)
+highp vec4 sdf_render(in highp float sdf, in highp vec4 sourceColor, in highp vec4 sdfColor, in highp float alpha, in highp float smoothing)
 {
     // This uses max to prevent a divide by zero when fwidth returns 0.0.
     highp float g = smoothing * max(fwidth(sdf), 0.0001);
@@ -227,7 +227,7 @@ lowp vec4 sdf_render(in lowp float sdf, in lowp vec4 sourceColor, in lowp vec4 s
 // \param sdfColor The color to use for rendering the sdf shape.
 //
 // \return sourceColor with the sdf shape rendered on top.
-lowp vec4 sdf_render(in lowp float sdf, in lowp vec4 sourceColor, in lowp vec4 sdfColor)
+highp vec4 sdf_render(in highp float sdf, in highp vec4 sourceColor, in highp vec4 sdfColor)
 {
     return sdf_render(sdf, sourceColor, sdfColor, 1.0, sdf_default_smoothing);
 }
@@ -239,7 +239,7 @@ lowp vec4 sdf_render(in lowp float sdf, in lowp vec4 sourceColor, in lowp vec4 s
 //
 // \param smoothing The amount of smoothing to apply to the sdf.
 //
-lowp vec4 sdf_render(in lowp float sdf, in lowp vec4 sourceColor, in lowp vec4 sdfColor, in lowp float smoothing)
+highp vec4 sdf_render(in highp float sdf, in highp vec4 sourceColor, in highp vec4 sdfColor, in highp float smoothing)
 {
     return sdf_render(sdf, sourceColor, sdfColor, 1.0, smoothing);
 }
